1. Added slug field in response classes including BusinessResponse, ChatSpaceResponse, PollResponse, SoftAskResponse, SoftAskReplyResponse and StreamResponse. 2. Updated BusinessMapperImpl, ChatSpaceMapperImpl, PollMapperImpl, StreamMapperImpl, SoftAskMapperImpl. 3. HasSlug to be implemented by response classes. 4. General refactoring.
