+++ src/main/java/com/fleencorp/feen/model/contract/HasSlug.java
@@ -1,0 +1,6 @@
+package com.fleencorp.feen.model.contract;
+
+public interface HasSlug {
+
+  void setSlug(String slug);
+}
