+++ src/main/java/com/fleencorp/feen/poll/service/PollVoteService.java
@@ -8,7 +8,7 @@ import com.fleencorp.feen.poll.exception.vote.PollVotingNotAllowedPollEndedExcep
 import com.fleencorp.feen.poll.exception.vote.PollVotingNotAllowedPollNoOptionException;
 import com.fleencorp.feen.poll.model.dto.VotePollDto;
 import com.fleencorp.feen.poll.model.request.PollVoteSearchRequest;
-import com.fleencorp.feen.poll.model.response.base.PollVoteResponse;
+import com.fleencorp.feen.poll.model.response.core.PollVoteResponse;
 import com.fleencorp.feen.poll.model.search.PollVoteSearchResult;
 import com.fleencorp.feen.user.exception.member.MemberNotFoundException;
 import com.fleencorp.feen.user.model.security.RegisteredUser;
