+++ src/main/java/com/fleencorp/feen/poll/service/impl/PollSearchServiceImpl.java
@@ -16,9 +16,9 @@ import com.fleencorp.feen.poll.model.info.PollVisibilityInfo;
 import com.fleencorp.feen.poll.model.request.PollSearchRequest;
 import com.fleencorp.feen.poll.model.response.GetDataRequiredToCreatePoll;
 import com.fleencorp.feen.poll.model.response.PollRetrieveResponse;
-import com.fleencorp.feen.poll.model.response.base.PollOptionResponse;
-import com.fleencorp.feen.poll.model.response.base.PollResponse;
-import com.fleencorp.feen.poll.model.response.base.PollVoteResponse;
+import com.fleencorp.feen.poll.model.response.core.PollOptionResponse;
+import com.fleencorp.feen.poll.model.response.core.PollResponse;
+import com.fleencorp.feen.poll.model.response.core.PollVoteResponse;
 import com.fleencorp.feen.poll.model.search.ChatSpacePollSearchResult;
 import com.fleencorp.feen.poll.model.search.PollSearchResult;
 import com.fleencorp.feen.poll.model.search.StreamPollSearchResult;
