+++ src/main/java/com/fleencorp/feen/poll/service/impl/PollVoteServiceImpl.java
@@ -14,8 +14,8 @@ import com.fleencorp.feen.poll.model.holder.PollVoteEntriesHolder;
 import com.fleencorp.feen.poll.model.info.IsVotedInfo;
 import com.fleencorp.feen.poll.model.info.TotalPollVoteEntriesInfo;
 import com.fleencorp.feen.poll.model.request.PollVoteSearchRequest;
-import com.fleencorp.feen.poll.model.response.base.PollOptionResponse;
-import com.fleencorp.feen.poll.model.response.base.PollVoteResponse;
+import com.fleencorp.feen.poll.model.response.core.PollOptionResponse;
+import com.fleencorp.feen.poll.model.response.core.PollVoteResponse;
 import com.fleencorp.feen.poll.model.search.PollVoteSearchResult;
 import com.fleencorp.feen.poll.service.PollCommonService;
 import com.fleencorp.feen.poll.service.PollOperationsService;
